try4

+++ main_server/src/main/java/ru/practicum/main_server/model/Event.java
@@ -34,7 +34,7 @@ public class Event {
     @Column(name = "paid", nullable = false)
     private boolean paid;
     @Column(name = "participant_limit")
-    private Integer participantLimit;
+    private Long participantLimit;
     @Column(name = "published_on")
     private LocalDateTime publishedOn;
     @Column(name = "request_moderation")
