+++ main_server/src/main/java/ru/practicum/main_server/model/dto/EventFullDto.java
@@ -19,7 +19,7 @@ public class EventFullDto {
     private UserShortDto initiator;
     private Location location;
     private boolean paid;
-    private int participantLimit;
+    private Long participantLimit;
     private String publishedOn;
     private boolean requestModeration;
     private String state;
